pyproject.toml
[project]
name = "strava-mcp-server"
version = "0.1.3"
description = "MCP server for Strava API integration"
readme = "README.md"
authors = [
    { name = "Tomek Korbak", email = "tomasz.korbak@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.1.0",
    "rich>=13.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
strava-mcp-server = "strava_mcp_server:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/tomek-korbak/strava-mcp-server"
"Bug Tracker" = "https://github.com/tomek-korbak/strava-mcp-server/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/strava_mcp_server"]

[tool.hatch.build]
only-packages = true

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "B", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true